+++ aa_forum/urls.py
@@ -63,6 +63,11 @@ urlpatterns = [
         admin.topic_change_lock_state,
         name="admin_topic_change_lock_state",
     ),
+    url(
+        r"^topic/(?P<topic_id>[0-9]+)/change-sticky-state/$",
+        admin.topic_change_sticky_state,
+        name="admin_topic_change_sticky_state",
+    ),
     # Forum URLs (This needs to be the last block!)
     url(
         r"^(?P<category_slug>[\w-]+)/(?P<board_slug>[\w-]+)/$",